backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
